+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   the last copy #3138
 - speed up loading of chat messages #3171
 - clear more columns when message expires due to `delete_device_after` setting #3181
+- do not try to use stale SMTP connections #3180
 
 ### Changes
 - add more SMTP logging #3093